requirements.txt
python-dotenv
langchain
langchain-community
langchain-groq
faiss-cpu
sentence-transformers
pymupdf
tiktoken